requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
